In short, this commit fixes our print-if problem.  Check below the fold to see what it actually does.
First, we disallow if-Exprs after lift-complex.  All ifs are Stmts,
and their results are set!'d to a variable so they may be used later.
However, since if-Exprs can return things of any type, the "result"
variable that is effected must also be of any type.  Since this
introduces the problem of initialization, we now allow a (let Type
Var) form that compiles to the C-style declaration "Type Var;".

In addition, remove-danger is moved way down in the compiler, but I
fear we must move it down even further... after infer-regions so that
our compiler is not slowed down in that bottleneck with much-bloated
code.